+++ builders/_service.js
@@ -42,6 +42,6 @@ module.exports = {
       // Add in creds if we have them
       if (options.creds) options.info.creds = options.creds;
       super(id, options, ...sources);
-    };
+    }
   },
 };