+++ examples/plugins/plugins/test-plugin/app.js
@@ -1,3 +1,3 @@
 'use strict';
 
-module.exports = injected => ({'app-plugin-test-app': true});
+module.exports = () => ({'app-plugin-test-app': true});