+++ hooks/app-purge-compose-dir.js
@@ -2,7 +2,7 @@
 
 const remove = require('../utils/remove');
 
-module.exports = async (app, lando) => {
+module.exports = async app => {
   // remove app compose directory and other things
   try {
     remove(app._dir);