+++ hooks/app-run-v4-destroy-service.js
@@ -25,14 +25,14 @@ module.exports = async (app, lando) => {
     const container = [app.project, service.id, '1'].join(lando.config.orchestratorSeparator);
     return {
       title: `Container ${container}`,
-      task: async (ctx, task) => {
+      task: async ctx => {
         try {
           await service.destroy();
         } catch (error) {
-          error = require('../utils/make-error')({error});
-          error.context = {id: container};
-          ctx.errors.push(error);
-          throw error;
+          const err = require('../utils/make-error')({error});
+          err.context = {id: container};
+          ctx.errors.push(err);
+          throw err;
         }
       },
     };