+++ hooks/lando-flush-updates-cache.js
@@ -1,6 +1,6 @@
 'use strict';
 
-module.exports = async (lando, options) => {
+module.exports = async lando => {
   if (lando.config.channel !== 'none' && lando.cache.get('updates-2')) {
     const {expires} = lando.cache.get('updates-2');
     if (expires < Date.now()) lando.cache.remove('updates-2');