+++ hooks/lando-setup-build-engine-darwin.js
@@ -60,7 +60,7 @@ const getEngineDownloadUrl = (id = '175267') => {
 /*
  * wrapper for docker-desktop install
  */
-const downloadDockerDesktop = (url, {debug, task, ctx}) => new Promise((resolve, reject) => {
+const downloadDockerDesktop = (url, {debug, task}) => new Promise((resolve, reject) => {
   const download = require('../utils/download-x')(url, {debug});
   // success
   download.on('done', result => {
@@ -129,45 +129,41 @@ module.exports = async (lando, options) => {
       return true;
     },
     task: async (ctx, task) => {
-      try {
-        // download the installer
-        ctx.download = await downloadDockerDesktop(url, {ctx, debug, task});
+      // download the installer
+      ctx.download = await downloadDockerDesktop(url, {ctx, debug, task});
 
-        // prompt for password if interactive
-        if (ctx.password === undefined && lando.config.isInteractive) {
-          ctx.password = await task.prompt({
-            type: 'password',
-            name: 'password',
-            message: `Enter computer password for ${lando.config.username} to install build engine`,
-            validate: async (input, state) => {
-              const options = {debug, ignoreReturnCode: true, password: input};
-              const response = await require('../utils/run-elevated')(['echo', 'hello there'], options);
-              if (response.code !== 0) return response.stderr;
-              return true;
-            },
-          });
-        }
-
-        task.title = `Installing build engine ${color.dim('(this may take a minute)')}`;
-
-        // assemble
-        const script = path.join(lando.config.userConfRoot, 'scripts', 'install-docker-desktop.sh');
-        const command = [script, '--installer', ctx.download.dest, '--user', lando.config.username];
-
-        // add optional args
-        if (options.buildEngineAcceptLicense) command.push('--accept-license');
-        if (options.debug || options.verbose > 0 || lando.debuggy) command.push('--debug');
-
-        // run
-        const result = await require('../utils/run-elevated')(command, {debug, password: ctx.password});
-        result.download = ctx.download;
-
-        // finish up
-        task.title = 'Installed build engine (Docker Desktop) to /Applications/Docker.app';
-        return result;
-      } catch (error) {
-        throw error;
+      // prompt for password if interactive
+      if (ctx.password === undefined && lando.config.isInteractive) {
+        ctx.password = await task.prompt({
+          type: 'password',
+          name: 'password',
+          message: `Enter computer password for ${lando.config.username} to install build engine`,
+          validate: async input => {
+            const options = {debug, ignoreReturnCode: true, password: input};
+            const response = await require('../utils/run-elevated')(['echo', 'hello there'], options);
+            if (response.code !== 0) return response.stderr;
+            return true;
+          },
+        });
       }
+
+      task.title = `Installing build engine ${color.dim('(this may take a minute)')}`;
+
+      // assemble
+      const script = path.join(lando.config.userConfRoot, 'scripts', 'install-docker-desktop.sh');
+      const command = [script, '--installer', ctx.download.dest, '--user', lando.config.username];
+
+      // add optional args
+      if (options.buildEngineAcceptLicense) command.push('--accept-license');
+      if (options.debug || options.verbose > 0 || lando.debuggy) command.push('--debug');
+
+      // run
+      const result = await require('../utils/run-elevated')(command, {debug, password: ctx.password});
+      result.download = ctx.download;
+
+      // finish up
+      task.title = 'Installed build engine (Docker Desktop) to /Applications/Docker.app';
+      return result;
     },
   });
 };