+++ plugins/proxy/builders/_proxy.js
@@ -78,6 +78,6 @@ module.exports = {
         userConfRoot: options.userConfRoot,
       };
       super('proxy', _.merge({}, config, augment), proxy, ports);
-    };
+    }
   },
 };