+++ utils/get-compose-bin-path.js
@@ -8,8 +8,9 @@ module.exports = () => {
       return '/Applications/Docker.app/Contents/Resources/bin';
     case 'linux':
       return require('./get-docker-bin-path')();
-    case 'win32':
+    case 'win32': {
       const programFiles = process.env.ProgramW6432 || process.env.ProgramFiles;
       return path.win32.join(programFiles + '\\Docker\\Docker\\resources\\bin');
+    }
   }
 };