+++ CHANGELOG.md
@@ -2,16 +2,16 @@
 
 ## v3.24.0-alpha.2 - [December 3, 2024](https://github.com/lando/core/releases/tag/v3.24.0-alpha.2)
 
-* This release has no new content it is just a test of dual branch developmebt and `edge` branch release deployment
+* This release has no new content it is just a test of dual branch development and `edge` branch release deployment
 
 ## v3.24.0-alpha.1 - [December 3, 2024](https://github.com/lando/core/releases/tag/v3.24.0-alpha.1)
 
-* This release has no new content it is just a test of dual branch developmebt and `edge` branch release deployment
+* This release has no new content it is just a test of dual branch development and `edge` branch release deployment
 
 ## v3.23.15 - [December 3, 2024](https://github.com/lando/core/releases/tag/v3.23.15)
 
 * Disabled `DOCKER_CLI_HINTS` on `lando`
-* Fixed bug causing some `lando exec` and tooling commands from not inheriting terminal columns and lines correctly[#277](https://github.com/lando/core/issues/277)
+* Fixed bug causing some `lando exec` and tooling commands from not inheriting terminal columns and lines correctly [#277](https://github.com/lando/core/issues/277)
 * Fixed bug causing `lando setup` loop on `U Need Setup!`
 * Fixed bug causing failed update message when user needs to relaunch terminal
 * Improved `ux` for `autosetup`